+++ Configuration/Settings.yaml
@@ -1,13 +1,13 @@
-TYPO3:
+Neos:
   Neos:
-    typoScript:
+    fusion:
       autoInclude:
-        'VIVOMEDIA.TypoScript.CacheVisualisation': TRUE
+        'VIVOMEDIA.Fusion.CacheVisualisation': true
 
 VIVOMEDIA:
-  TypoScript:
+  Fusion:
     CacheVisualisation:
-      enabled: FALSE
+      enabled: false
       output:
         wrapperClassName: 'vivomedia-cachevisualisation'
         itemClassName:
@@ -15,7 +15,7 @@ VIVOMEDIA:
           uncached: 'vivomedia-cachevisualisation-uncached'
       pathBlacklist:
         - 'root'
-        - 'root<TYPO3.TypoScript:Case>/default'
-        - 'root<TYPO3.TypoScript:Case>/default<TYPO3.TypoScript:Matcher>/element<TYPO3.Neos.NodeTypes:Page.Document>'
-        - 'root<TYPO3.TypoScript:Case>/default<TYPO3.TypoScript:Matcher>/element<TYPO3.Neos.NodeTypes:Page.Document>/head<TYPO3.TypoScript:Array>/neosBackendEndpoints'
-        - 'root<TYPO3.TypoScript:Case>/default<TYPO3.TypoScript:Matcher>/element<TYPO3.Neos.NodeTypes:Page.Document>/head<TYPO3.TypoScript:Array>/neosBackendHeader'
+        - 'root<Neos.Fusion:Case>/default'
+        - 'root<Neos.Fusion:Case>/default<Neos.Fusion:Matcher>/element<Neos.NodeTypes:Page.Document>'
+        - 'root<Neos.Fusion:Case>/default<Neos.Fusion:Matcher>/element<Neos.NodeTypes:Page.Document>/head<Neos.Fusion:Array>/neosBackendEndpoints'
+        - 'root<Neos.Fusion:Case>/default<Neos.Fusion:Matcher>/element<Neos.NodeTypes:Page.Document>/head<Neos.Fusion:Array>/neosBackendHeader'
