--- Resources/Private/TypoScript/Root.ts2
@@ -1,25 +1,0 @@
-prototype(TYPO3.Neos:Page){
-  head {
-    cacheVisualisationCss = TYPO3.TypoScript:Tag {
-      tagName = 'link'
-      attributes {
-        rel = ${'stylesheet'}
-        href = TYPO3.TypoScript:ResourceUri {
-          path = 'resource://VIVOMEDIA.TypoScript.CacheVisualisation/Public/css/cachevisualisation.css'
-        }
-      }
-      @if.onlyRenderNotInBackend = ${!node.context.inBackend}
-    }
-  }
-
-  cacheVisualisationJs = TYPO3.TypoScript:Tag {
-    @position = 'before closingBodyTag'
-    tagName = 'script'
-    attributes {
-      src = TYPO3.TypoScript:ResourceUri {
-        path = 'resource://VIVOMEDIA.TypoScript.CacheVisualisation/Public/js/cachevisualisation.js'
-      }
-    }
-    @if.onlyRenderNotInBackend = ${!node.context.inBackend}
-  }
-}
